[ADD] - Add function send_command for all commands when send in server
