+++ client/include/get_command.h
@@ -60,6 +60,8 @@ int check_params(char *av);
 
 char **parse_response(char *response, int nb_params);
 
+char **send_command(char *av, char **tab_response, char *command, int socket);
+
 int parse_command(char *input, int socket);
 
 #endif /* !GET_COMMAND_H_ */
