+++ client/src/manage_commands/send_command.c
@@ -1,0 +1,31 @@
+/*
+** EPITECH PROJECT, 2022
+** Epi-MyTeams
+** File description:
+** send_command
+*/
+#include "client.h"
+#include "get_command.h"
+#include "reply_codes.h"
+#include "logging_client.h"
+
+char **send_command(char *av, char **tab_response, char *command, int socket)
+{
+    char buff[4096];
+    char response[2];
+
+    memset(buff, 0, 4096);
+    make_command_rfc_compatible(buff, command, av);
+    write(socket, buff, strlen(buff));
+    memset(buff, 0, 4096);
+    if (read(socket, buff, 4096) == 0) {
+        printf("Client is deconnected !\n");
+        exit (0);
+    }
+    client_reply(atoi(strncpy(response, buff, 2)));
+    if (strlen(buff) > 4) {
+        tab_response = parse_response(buff, 2);
+        client_event_logged_in(tab_response[1], tab_response[3]);
+    }
+    return tab_response;
+}
